requirements.txt
google-generativeai
pyaudio
python-dotenv
